Documentation update. Name update
