+++ docs/Advanced/converter.rst
@@ -1,0 +1,51 @@
+.. _convert:
+
+==========
+Converters
+==========
+
+#############
+To continuous
+#############
+
+.. automodule:: zellij.utils.converters
+  :members: Continuous
+  :undoc-members:
+  :show-inheritance:
+
+******
+MinMax
+******
+
+.. automodule:: zellij.utils.converters
+   :members: ArrayMinmax, FloatMinmax, IntMinmax, CatMinmax, ConstantMinmax
+   :undoc-members:
+   :show-inheritance:
+
+###########
+To discrete
+###########
+
+.. automodule:: zellij.utils.converters
+  :members: ArrayBinning, FloatBinning, IntBinning, CatBinning, ConstantBinning
+  :undoc-members:
+  :show-inheritance:
+
+*******
+Binning
+*******
+
+.. automodule:: zellij.utils.converters
+  :members: Discrete
+  :undoc-members:
+  :show-inheritance:
+
+
+#######
+Others
+#######
+
+.. automodule:: zellij.utils.converters
+  :members: DoNothing
+  :undoc-members:
+  :show-inheritance:
