+++ docs/Advanced/distance.rst
@@ -1,0 +1,10 @@
+.. _dist:
+
+=========
+Distances
+=========
+
+.. automodule:: zellij.utils.distances
+  :members:
+  :undoc-members:
+  :show-inheritance:
