+++ docs/Advanced/index.rst
@@ -1,3 +1,10 @@
 ==============
 Other features
 ==============
+
+.. toctree::
+
+   addons
+   converter
+   neighborhood
+   distance
