+++ docs/Advanced/neighborhood.rst
@@ -4,4 +4,22 @@
 Neighborhood
 ============
 
-Coming soon
+Neighborhood objets are :ref:`addons` defining what a neighborhood is for
+a :ref:`var` via :ref:`varadd`, or a :ref:`sp` via :ref:`spadd`.
+
+#########
+Intervals
+#########
+
+Intervals :ref:`nbh` are based on drawing a random value between :math:`x \pm neighborhood`.
+
+.. automodule:: zellij.utils.neighborhoods
+   :members: Intervals
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: zellij.utils.neighborhoods
+   :members: ArrayInterval, FloatInterval, IntInterval, CatInterval, ConstantInterval
+   :undoc-members:
+   :show-inheritance:
