+++ docs/Core/index.rst
@@ -4,7 +4,7 @@ Main objects
 
 .. toctree::
 
+   variables
    searchspace
    lossfunc
    metaheuristic
-   neighborhood
