+++ docs/Core/searchspace.rst
@@ -1,10 +1,14 @@
 .. _sp:
 
 ============
-Search Space
+Search space
 ============
 
+The :ref:`sp` will merge the :ref:`lf` and defined :ref:`var`.
+:ref:`sp` will alows to draw random points and attributes.
+:ref:`sp` features can be extended by grafting :ref:`addons` to it.
+
 .. automodule:: zellij.core.search_space
-   :members: Searchspace
+   :members: Searchspace, MixedSearchspace, ContinuousSearchspace, DiscreteSearchspace
    :undoc-members:
    :show-inheritance:
