+++ docs/Dac/dac.rst
@@ -1,0 +1,10 @@
+.. _dac:
+
+============================
+Divide-and-Conquer algorithm
+============================
+
+.. automodule:: zellij.strategies.dac
+   :members:
+   :undoc-members:
+   :show-inheritance:
