+++ docs/Dac/geometry.rst
@@ -1,0 +1,39 @@
+.. _frac:
+
+========
+Geometry
+========
+
+
+This module allows to define search space based on an Euclidean geometrical object.
+It works the same as :ref:`sp`, but all dimensions are continuous.
+
+*******
+Classic
+*******
+
+.. automodule:: zellij.core.geometry
+   :members: Fractal, Hypercube, Hypersphere, Section
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+*************
+Voronoï based
+*************
+
+.. automodule:: zellij.core.geometry
+  :members: Voronoi, DynamicVoronoi, FixedVoronoi, LightFixedVoronoi, BoxedVoronoi
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
+
+*************
+Miscellaneous
+*************
+
+.. automodule:: zellij.core.geometry
+  :members: Direct, Soo
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
