+++ docs/Dac/heuristics.rst
@@ -1,0 +1,42 @@
+.. _heuristic:
+
+===============
+Scoring methods
+===============
+
+.. automodule:: zellij.strategies.utils.heuristics
+   :members: Heuristic
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+
+******
+Basics
+******
+
+.. automodule:: zellij.strategies.utils.heuristics
+   :members: Min, Median, Mean, Std
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+*************
+Fuzzy measure
+*************
+
+.. automodule:: zellij.strategies.utils.heuristics
+  :members: Belief
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
+
+*************
+Miscellaneous
+*************
+
+.. automodule:: zellij.strategies.utils.heuristics
+  :members: Distance_to_the_best, Distance_to_the_best_corrected
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
