+++ docs/Dac/index.rst
@@ -1,0 +1,10 @@
+=======================
+Decomposition algorithm
+=======================
+
+.. toctree::
+
+   dac
+   geometry
+   tree_search
+   heuristics
