+++ docs/Dac/tree_search.rst
@@ -1,0 +1,52 @@
+.. _tree_search:
+
+======================
+Tree search algorithms
+======================
+
+.. automodule:: zellij.strategies.utils.tree_search
+   :members: Tree_search
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+
+******
+Basics
+******
+
+.. automodule:: zellij.strategies.utils.tree_search
+   :members: Breadth_first_search, Depth_first_search, Best_first_search
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+********
+Advanced
+********
+
+.. automodule:: zellij.strategies.utils.tree_search
+   :members: Epsilon_greedy_search, Diverse_best_first_search
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+********
+Prunning
+********
+
+.. automodule:: zellij.strategies.utils.tree_search
+  :members: Beam_search, Cyclic_best_first_search
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
+
+*************
+Miscellaneous
+*************
+
+.. automodule:: zellij.strategies.utils.tree_search
+  :members: Potentially_Optimal_Rectangle, Locally_biased_POR, Adaptive_POR, Soo_tree_search, Move_up
+  :undoc-members:
+  :show-inheritance:
+  :noindex:
