--- docs/Fda/index.rst
@@ -1,32 +1,0 @@
-.. _fda:
-
-===============================
-Fractal Decomposition Algorithm
-===============================
-
-.. automodule:: zellij.strategies.fda
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. _frac:
-
-********
-Fractals
-********
-
-Classic
-#######
-
-.. automodule:: zellij.core.fractals
-   :members: Fractal, Hypercube, Hypersphere, Section
-   :undoc-members:
-   :show-inheritance:
-
-Voronoï based
-#############
-
-.. automodule:: zellij.core.fractals
-  :members: Voronoi, DynamicVoronoi, FixedVoronoi, LightFixedVoronoi, BoxedVoronoi
-  :undoc-members:
-  :show-inheritance:
