+++ docs/Strategies/ga_base.rst
@@ -8,3 +8,37 @@ Genetic Algorithm
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. _gaadd:
+
+########################
+Genetic Algorithm Addons
+########################
+
+Here is the list of :ref:`Addons` linked to :ref:`ga`.
+These are implemented by using :ref:`sp` kwargs.
+
+********
+Mutation
+********
+
+.. automodule:: zellij.utils.operators
+   :members: NeighborMutation
+   :undoc-members:
+   :show-inheritance:
+
+*********
+Crossover
+*********
+.. automodule:: zellij.utils.operators
+   :members: DeapOnePoint
+   :undoc-members:
+   :show-inheritance:
+
+*********
+Selection
+*********
+.. automodule:: zellij.utils.operators
+   :members: DeapTournament
+   :undoc-members:
+   :show-inheritance:
