+++ lib/zellij/core/addons.py
@@ -16,17 +16,19 @@ logger = logging.getLogger("zellij.addons")
 class Addon(ABC):
     """Addon
 
-    An `Addon` in Zellij, is an additionnal feature that can be added to a `target` object.
-    See `VarAddon` for addon targeting `Variables` or `SearchspaceAddon` targeting :ref:`sp`.
+    Abstract class describing what an addon is.
+    An `Addon` in Zellij, is an additionnal feature that can be added to a
+    `target` object. See :ref:`varadd` for addon targeting :ref:`var` or
+    :ref:`spadd` targeting :ref:`sp`.
 
     Parameters
     ----------
-    object : object
+    target : Object, default=None
         Object targeted by the addons
 
     Attributes
     ----------
-    target : object
+    target : Object, default=None
         Object targeted by the addons
 
     """
@@ -44,6 +46,22 @@ class Addon(ABC):
 
 
 class VarAddon(Addon):
+    """VarAddon
+
+    :ref:`addons` where the target must be of type :ref:`var`.
+
+    Parameters
+    ----------
+    target : :ref:`var`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`var`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, variable=None):
         super(VarAddon, self).__init__(variable)
 
@@ -64,6 +82,22 @@ class VarAddon(Addon):
 
 
 class SearchspaceAddon(Addon):
+    """SearchspaceAddon
+
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(SearchspaceAddon, self).__init__(search_space)
 
@@ -85,6 +119,23 @@ class SearchspaceAddon(Addon):
 
 
 class Neighborhood(SearchspaceAddon):
+    """Neighborhood
+
+    :ref:`addons` where the target must be of type :ref:`sp`.
+    Describes what a neighborhood is for a :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, neighborhood, search_space=None):
         super(Neighborhood, self).__init__(search_space)
         self.neighborhood = neighborhood
@@ -99,6 +150,23 @@ class Neighborhood(SearchspaceAddon):
 
 
 class VarNeighborhood(VarAddon):
+    """VarNeighborhood
+
+    :ref:`addons` where the target must be of type :ref:`var`.
+    Describes what a neighborhood is for a :ref:`var`.
+
+    Parameters
+    ----------
+    target : :ref:`var`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`var`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, neighborhood, variable=None):
         super(VarAddon, self).__init__(variable)
         self.neighborhood = neighborhood
@@ -113,6 +181,25 @@ class VarNeighborhood(VarAddon):
 
 
 class Converter(SearchspaceAddon):
+    """Converter
+
+    :ref:`addons` where the target must be of type :ref:`sp`.
+    Describes what a converter is for a :ref:`sp`.
+    Converter allows to convert the type of a :ref:`sp` to another one.
+    All :ref:`var` must have a converter :ref:`varadd` implemented.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(Converter, self).__init__(search_space)
 
@@ -126,6 +213,24 @@ class Converter(SearchspaceAddon):
 
 
 class VarConverter(VarAddon):
+    """VarConverter
+
+    :ref:`addons` where the target must be of type :ref:`var`.
+    Describes what a converter is for a :ref:`var`.
+    Converter allows to convert the type of a :ref:`var` to another one.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, variable=None):
         super(VarConverter, self).__init__(variable)
 
@@ -139,6 +244,23 @@ class VarConverter(VarAddon):
 
 
 class Operator(SearchspaceAddon):
+    """Operator
+
+    Abstract class describing what an operator is for a :ref:`sp`.
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(Operator, self).__init__(search_space)
 
@@ -148,21 +270,89 @@ class Operator(SearchspaceAddon):
 
 
 class Mutator(SearchspaceAddon):
+    """Mutator
+
+    Abstract class describing what an Mutator is for a :ref:`sp`.
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(Mutator, self).__init__(search_space)
 
 
 class Crossover(SearchspaceAddon):
+    """Crossover
+
+    Abstract class describing what an MCrossover is for a :ref:`sp`.
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(Crossover, self).__init__(search_space)
 
 
 class Selector(SearchspaceAddon):
+    """Selector
+
+    Abstract class describing what an Selector is for a :ref:`sp`.
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None):
         super(Selector, self).__init__(search_space)
 
 
 class Distance(SearchspaceAddon):
+    """Distance
+
+    Abstract class describing what an Distance is for a :ref:`sp`.
+    :ref:`addons` where the target must be of type :ref:`sp`.
+
+    Parameters
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    Attributes
+    ----------
+    target : :ref:`sp`, default=None
+        Object targeted by the addons
+
+    """
+
     def __init__(self, search_space=None, weights=None):
         super(Distance, self).__init__(search_space)
         self.weights = None
